fixed switched layouts for the accounts subdirectory layout and the root layout
